+++ tsetmc/__init__.py
@@ -14,7 +14,7 @@ from pandas import read_csv as _read_csv, DataFrame as _DataFrame
 # noinspection PyUnresolvedReferences
 from pandas import to_numeric as _to_numeric, read_html as _read_html
 from aiohttp import ClientSession as _ClientSession, \
-    ClientTimeout as _ClientTimeout
+    ClientTimeout as _ClientTimeout, TCPConnector as _TCPConnector
 
 
 _csv2df = _partial(_read_csv, low_memory=False, engine='c', lineterminator=';')
@@ -103,10 +103,19 @@ SESSION : _ClientSession | None = None
 
 class Session(_ClientSession):
 
-    def __init__(self, **kwargs):
+    def __init__(self, /, **kwargs):
+        """Create a ClientSession object.
+
+        Use
+        ``ClientTimeout(total=30., sock_connect=5., sock_read=5.)``
+        as the default timeout and
+        ``TCPConnector(limit_per_host=1, keepalive_timeout=120.)``
+        as the default connector.
+        """
         if 'timeout' not in kwargs:
-            kwargs['timeout'] = _ClientTimeout(
-                total=30, sock_connect=5, sock_read=5)
+            kwargs['timeout'] = _ClientTimeout(total=30., sock_connect=5., sock_read=5.)
+        if 'connector' not in kwargs:
+            kwargs['connector'] = _TCPConnector(limit_per_host=1, keepalive_timeout=120.)
         super().__init__(**kwargs)
 
     async def __aenter__(self) -> _ClientSession: